backend/requirements.txt
fastapi==0.110.3
uvicorn[standard]==0.29.0
sqlalchemy==2.0.49
pydantic==2.8.2
pydantic-core==2.20.1
python-jose[cryptography]==3.3.0
passlib==1.7.4
python-dotenv==1.0.1
python-multipart==0.0.9
alembic==1.13.1
email-validator==2.1.1
psycopg2-binary==2.9.11
openai==1.30.1
google-generativeai==0.8.3
slowapi
requests
razorpay==1.4.2
